refactor: rename `colorfy` module to `theme` for clarity
